+++ inst/d3js/modelStudio.js
@@ -34,6 +34,7 @@ var TIME = options.time,
     SHOW_SUBTITLE = options.show_subtitle,
     subTitle = options.subtitle || modelName,
     msTitle = options.ms_title,
+    msSubtitle = options.ms_subtitle,
     barWidth = options.bar_width,
     lineSize = options.line_size,
     pointSize = options.point_size,
@@ -98,7 +99,8 @@ var TIME = options.time,
     atBarColor = options.at_bar_color || barColor,
     atLineColor = options.at_line_color || lineColor,
     atPointColor = options.at_point_color || pointColor,
-    telemetry = options.telemetry;
+    telemetry = options.telemetry,
+    license = options.license;
 
 /// for observation choice
 var observationIds = Object.keys(obsData);
@@ -120,7 +122,7 @@ var w = options.w, h = options.h;
 var plotWidth = w + margin.left + margin.right,
     plotHeight = h + margin.top + margin.bottom;
 
-var studioMargin = {top: 50, bottom: 50},
+var studioMargin = {top: options.ms_margin_top, bottom: options.ms_margin_bottom},
     studioWidth = dim[1]*plotWidth,
     studioHeight = studioMargin.top + dim[0]*plotHeight + studioMargin.bottom;
 
@@ -173,6 +175,7 @@ for (let i = 0; i < dim[0]; i++) {
 }
 
 ///:\\\
+if (license) document.head.appendChild(document.createComment(license));
 if (telemetry) startTelemetrySession();
 initializeStudio();
 ///:\\\
@@ -196,6 +199,15 @@ function initializeStudio() {
        .attr("x", 15)
        .attr("y", 30)
        .text(msTitle);
+
+  if (msSubtitle !== null) {
+      TOP_G.append("text")
+           .attr("class", "subTitle")
+           .attr("x", 15)
+           .attr("y", 40)
+           .text(msSubtitle)
+           .call(wrapText, studioWidth - 15);
+  }
 
   TOP_G.append("line")
        .attr("class", "mainLine")